Moving activate/deactivate into the Process class and adapting callers.
Note that this change is similar to the changes earlier to the BaseActivity
classes. The difference is that Process is handled directly, where as activities
are coming in thru a factory in the manager. Both ''should'' be the same, but
now we can compare the two to see which one works best. Changing the one into 
the other is not much work.


	* lib/Galaxia/API.php: ...
	* lib/Galaxia/src/API/Process.php: ...
	* lib/Galaxia/src/ProcessManager/ActivityManager.php: ...
	* lib/Galaxia/src/ProcessManager/ProcessManager.php: ...
	* xaradmin/activities.php: ...
	* xaradmin/graph.php: ...
	* xaradmin/processes.php: ...