Bug 6322 - Update existing xarbb hooks with ability to create a link in the forum topic to the hooked article, and vice versa so a link goes in the article to the hooked forum topic.
At the moment the relevant forum must have bbcode or html checked in the forum option to enable the options for this xarbb hook. Perhaps a text link could be added, in the future, for forums without html or bbcode but at the momemnt this does not seem that useful.

	* xaradmin/modify.php: ...
	* xaradmin/modifyconfig.php: ...
	* xaradmin/newhook.php: ...
	* xaradminapi/createhook.php: ...
	* xartemplates/admin-modify.xd: ...
	* xartemplates/admin-modifyconfig.xd: ...

applied changes from 3e90ba6cfed7b0d4171bd962124b2507242f5bf9
             through 677b5e026ef02e7e3388b73fefeb975ab2270cd0